feat: add TestUserUtils
