+++ portal/src/it/scala/cromwell/pipeline/controller/AuthControllerItTest.scala
@@ -1,16 +1,14 @@
 package cromwell.pipeline.controller
 
-import java.util.UUID
-
 import akka.http.scaladsl.model.ContentTypes.`application/json`
 import akka.http.scaladsl.model.{ HttpEntity, StatusCodes }
 import akka.http.scaladsl.testkit.ScalatestRouteTest
 import com.dimafeng.testcontainers.{ ForAllTestContainer, PostgreSQLContainer }
 import com.typesafe.config.Config
 import cromwell.pipeline.controller.AuthController._
+import cromwell.pipeline.datastorage.dto.User
 import cromwell.pipeline.datastorage.dto.auth.{ SignInRequest, SignUpRequest }
-import cromwell.pipeline.datastorage.dto.{ User, UserId }
-import cromwell.pipeline.utils.StringUtils
+import cromwell.pipeline.utils.auth.TestUserUtils
 import cromwell.pipeline.{ ApplicationComponents, TestContainersUtils }
 import org.scalatest.compatible.Assertion
 import org.scalatest.concurrent.ScalaFutures._
@@ -27,8 +25,8 @@ class AuthControllerItTest extends WordSpec with Matchers with ScalatestRouteTes
   override protected def beforeAll(): Unit =
     components.datastorageModule.pipelineDatabaseEngine.updateSchema()
 
-  import components.datastorageModule.userRepository
   import components.controllerModule.authController
+  import components.datastorageModule.userRepository
 
   private val userPassword = "-Pa$$w0rd-"
 
@@ -37,10 +35,10 @@ class AuthControllerItTest extends WordSpec with Matchers with ScalatestRouteTes
     "signIn" should {
 
       "return token headers if user exists" in {
-        val newUser = getDummyUser(userPassword)
+        val dummyUser: User = TestUserUtils.getDummyUser()
 
-        whenReady(userRepository.addUser(newUser)) { _ =>
-          val signInRequest = SignInRequest(newUser.email, userPassword)
+        whenReady(userRepository.addUser(dummyUser)) { _ =>
+          val signInRequest = SignInRequest(dummyUser.email, userPassword)
           val httpEntity = HttpEntity(`application/json`, Json.stringify(Json.toJson(signInRequest)))
 
           Post("/auth/signIn", httpEntity) ~> authController.route ~> check {
@@ -54,8 +52,8 @@ class AuthControllerItTest extends WordSpec with Matchers with ScalatestRouteTes
     "signUp" should {
 
       "return token headers if user was successfully registered" in {
-        val newUser = getDummyUser(userPassword)
-        val signUpRequest = SignUpRequest(newUser.email, userPassword, newUser.firstName, newUser.lastName)
+        val dummyUser: User = TestUserUtils.getDummyUser()
+        val signUpRequest = SignUpRequest(dummyUser.email, userPassword, dummyUser.firstName, dummyUser.lastName)
         val httpEntity = HttpEntity(`application/json`, Json.stringify(Json.toJson(signUpRequest)))
 
         Post("/auth/signUp", httpEntity) ~> authController.route ~> check {
@@ -68,10 +66,10 @@ class AuthControllerItTest extends WordSpec with Matchers with ScalatestRouteTes
     "refresh" should {
 
       "return updated token headers if refresh token was valid and active" in {
-        val newUser = getDummyUser(userPassword)
+        val dummyUser: User = TestUserUtils.getDummyUser()
 
-        whenReady(userRepository.addUser(newUser)) { _ =>
-          val signInRequest = SignInRequest(newUser.email, userPassword)
+        whenReady(userRepository.addUser(dummyUser)) { _ =>
+          val signInRequest = SignInRequest(dummyUser.email, userPassword)
           val httpEntity = HttpEntity(`application/json`, Json.stringify(Json.toJson(signInRequest)))
 
           Post("/auth/signIn", httpEntity) ~> authController.route ~> check {
@@ -86,20 +84,6 @@ class AuthControllerItTest extends WordSpec with Matchers with ScalatestRouteTes
       }
     }
 
-  }
-
-  private def getDummyUser(password: String = userPassword, passwordSalt: String = "salt"): User = {
-    val uuid = UUID.randomUUID().toString
-    val passwordHash = StringUtils.calculatePasswordHash(password, passwordSalt)
-    User(
-      userId = UserId(uuid),
-      email = s"JohnDoe-$uuid@cromwell.com",
-      passwordHash = passwordHash,
-      passwordSalt = passwordSalt,
-      firstName = "FirstName",
-      lastName = "LastName",
-      profilePicture = None
-    )
   }
 
   private def checkAuthTokens: Assertion =
