+++ portal/src/it/scala/cromwell/pipeline/controller/UserControllerItTest.scala
@@ -1,19 +1,17 @@
 package cromwell.pipeline.controller
 
-import java.util.UUID
-
 import akka.http.scaladsl.model.StatusCodes
 import akka.http.scaladsl.testkit.ScalatestRouteTest
 import com.dimafeng.testcontainers.{ ForAllTestContainer, PostgreSQLContainer }
 import com.typesafe.config.Config
-import cromwell.pipeline.datastorage.dto.{ User, UserDeactivationByEmailResponse, UserDeactivationByIdResponse, UserId }
-import cromwell.pipeline.utils.StringUtils
+import cromwell.pipeline.datastorage.dto.{ User, UserDeactivationByEmailResponse, UserDeactivationByIdResponse }
+import cromwell.pipeline.utils.auth.TestUserUtils
 import cromwell.pipeline.{ ApplicationComponents, TestContainersUtils }
 import de.heikoseeberger.akkahttpplayjson.PlayJsonSupport._
 import org.scalatest.concurrent.ScalaFutures.whenReady
-import org.scalatest.{ AsyncWordSpec, Matchers }
+import org.scalatest.{ Matchers, WordSpec }
 
-class UserControllerItTest extends AsyncWordSpec with Matchers with ScalatestRouteTest with ForAllTestContainer {
+class UserControllerItTest extends WordSpec with Matchers with ScalatestRouteTest with ForAllTestContainer {
   override val container: PostgreSQLContainer = TestContainersUtils.getPostgreSQLContainer()
   container.start()
   implicit val config: Config = TestContainersUtils.getConfigForPgContainer(container)
@@ -25,15 +23,13 @@ class UserControllerItTest extends AsyncWordSpec with Matchers with ScalatestRou
   import components.controllerModule.userController
   import components.datastorageModule.userRepository
 
-  private val userPassword = "-Pa$$w0rd-"
-
   "UserController" when {
     "deactivateByEmail" should {
       "return email and false value if user was successfully deactivated" in {
-        val newUser = getDummyUser(userPassword)
-        val emailResponse = UserDeactivationByEmailResponse(newUser.email, false)
-        whenReady(userRepository.addUser(newUser)) { _ =>
-          val deactivateUserByEmailRequest = newUser.email
+        val dummyUser: User = TestUserUtils.getDummyUser()
+        val emailResponse = UserDeactivationByEmailResponse(dummyUser.email, false)
+        whenReady(userRepository.addUser(dummyUser)) { _ =>
+          val deactivateUserByEmailRequest = dummyUser.email
           Delete("/users/deactivate", deactivateUserByEmailRequest) ~> userController.route ~> check {
             responseAs[UserDeactivationByEmailResponse] shouldBe emailResponse
             status shouldBe StatusCodes.OK
@@ -43,28 +39,15 @@ class UserControllerItTest extends AsyncWordSpec with Matchers with ScalatestRou
     }
     "deactivateById" should {
       "return id and false value if user was successfully deactivated" in {
-        val newUser = getDummyUser(userPassword)
-        val idResponse = UserDeactivationByIdResponse(newUser.userId, false)
-        whenReady(userRepository.addUser(newUser)) { _ =>
-          Delete(s"/users/deactivate/${newUser.userId.value}") ~> userController.route ~> check {
+        val dummyUser: User = TestUserUtils.getDummyUser()
+        val idResponse = UserDeactivationByIdResponse(dummyUser.userId, false)
+        whenReady(userRepository.addUser(dummyUser)) { _ =>
+          Delete(s"/users/deactivate/${dummyUser.userId.value}") ~> userController.route ~> check {
             responseAs[UserDeactivationByIdResponse] shouldBe idResponse
             status shouldBe StatusCodes.OK
           }
         }
       }
     }
-  }
-  private def getDummyUser(password: String = userPassword, passwordSalt: String = "salt"): User = {
-    val uuid = UUID.randomUUID().toString
-    val passwordHash = StringUtils.calculatePasswordHash(password, passwordSalt)
-    User(
-      userId = UserId(uuid),
-      email = s"JohnDoe-$uuid@cromwell.com",
-      passwordHash = passwordHash,
-      passwordSalt = passwordSalt,
-      firstName = "FirstName",
-      lastName = "LastName",
-      profilePicture = None
-    )
   }
 }
