+++ portal/src/main/scala/cromwell/pipeline/controller/UserController.scala
@@ -37,4 +37,5 @@ class UserController(userService: UserService)(implicit executionContext: Execut
       }
     )
   }
+
 }
